Use "unsigned long long int" for NameGenerator count, to make absolutely sure it will never create duplicates in any realistic amount of time.

+++ OgreMain/include/OgreCommon.h
@@ -677,7 +677,7 @@ namespace Ogre {
 	{
 	protected:
 		String mPrefix;
-		unsigned long mNext;
+		unsigned long long int mNext;
 		OGRE_AUTO_MUTEX
 	public:
 		NameGenerator(const String& prefix) : mPrefix(prefix), mNext(1) {}
